Room and Agent Changes
Changed how rooms operate. Now a room is an instance, but can have a
special parent type of a room. Will show up as a room if we ask if it is
because it's parent is.

Agent actionary name is no longer hard-coded. But there needs to be an
agent in the actionary, and then agents that do not have a parent name
match will match to the first agent in the actionary.